# trabalho/pa/a9_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW Re-Event Web API'
  description: 'Web Resources Specification (A9) for Re-Event'
servers:
- url: http://lbaw2291.lbaw.fe.up.pt
  description: Production server
...
paths:
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Form page that allows user login. Access: PUB'
      tags:
        - 'M01: Authentication and User Profile'
      responses:
        '200':
          description: 'Ok. Show log-in UI'
    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Web resource that verifies user input in order to perform authentication. Access: PUB'
      tags:
        - 'M01: Authentication and User Profile'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:          # <!--- form field name
                  type: string
                password:    # <!--- form field name
                  type: string
              required:
                - email
                - password

      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to home page.'
                  value: '/home'
                302Error:
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
...
  /register:
    get:
      operationId: R104
      summary: 'R104: Register Form'
      description: 'Form page that allows user to register themselves into the application. Access: PUB'
      tags:
        - 'M01: Authentication and User Profile'
      responses:
        '200':
          description: 'Ok. Show register UI'

    post:
      operationId: R105
      summary: 'R105: Register Action'
      description: 'This web resource register the user into the system, processing the new user registration form submission. Access: PUB'
      tags:
        - 'M01: Authentication and User Profile'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                birthdate:
                  type: string
                  format: date
                gender:
                  type: string
                password:
                  type: string
                password_confirmation:
                  type: string
                picture:
                  type: string
                  format: binary
              required:
                - name
                - email
                - birthdate
                - password
                - password_confirmation

      responses:
        '302':
          description: 'Redirect after processing the new user information.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to home page.'
                  value: '/home'
                302Failure:
                  description: 'Failed authentication. Redirect to register form.'
                  value: '/register'
...
  /deleteUser:

    post:
        operationId: R108
        summary: 'R108: Delete User Action'
        description: 'This web resource allows the user to delete their own profile or admin to delete user profiles. Access: AUTH, OWN, ADM'
        tags:
          - 'M01: Authentication and User Profile'

        requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  name:
                    type: integer
                required:
                  - userid

        responses:
          '302':
            description: 'Redirect after processing the delete user information.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  302Success:
                    description: 'Successful delete. Redirect to home page (now unauthenticated).'
                    value: '/home'
                  302Failure:
                    description: 'Failed delete. Redirect to same user page.'
                    value: '/user{id}'
          '400':
            description: 'Bad Request.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  400:
                    description: 'Bad request. Redirect to the same user page'
                    value: '/user{id}'      
  /forgot-password:

    get:
      operationId: R109
      summary: 'R109: Forgot Password Form'
      description: 'Form page that allows user to ask for a forgot password email. Access: PUB'
      tags:
        - 'M01: Authentication and User Profile'
      responses:
        '200':
          description: 'Ok. Show forgot password UI'
    post:
        operationId: R110
        summary: 'R110: Forgot Password Action'
        description: 'This web resource allows the user to ask for an email to reset their password. Access: AUTH, OWN'
        tags:
          - 'M01: Authentication and User Profile'

        requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  name:
                    type: integer
                required:
                  - userid

        responses:
          '302':
            description: 'Redirect after the successful email sent.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  302Success:
                    description: 'Successful reset email sent. Redirect to home page (now unauthenticated).'
                    value: '/home'
                  302Failure:
                    description: 'Unsuccessful reset email sent. Redirect to the same form.'
                    value: '/forgot-password'
          '400':
            description: 'Bad Request.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  400:
                    description: 'Bad request. Redirect to the same form.'
                    value: '/forgot-password'                
  /reset-password:

    get:
      operationId: R111
      summary: 'R111: Reset Password Form'
      description: 'Form page that allows user to reset their password. Access: OWN'
      tags:
        - 'M01: Authentication and User Profile'
      responses:
        '200':
          description: 'Ok. Show reset password password UI'

    post:
        operationId: R112
        summary: 'R112: Reset Password Action'
        description: 'This web resource allows the user to reset their password on the system. Access: AUTH, OWN'
        tags:
          - 'M01: Authentication and User Profile'

        requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  name:
                    type: integer
                required:
                  - userid

        responses:
          '302':
            description: 'Redirect after resetting th user password.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  302Success:
                    description: 'Successful reset. Redirect to home page (now unauthenticated).'
                    value: '/home'
                  302Failure:
                    description: 'Unsuccessful reset. Redirect to the same form.'
                    value: '/reset-password{token}'
          '400':
            description: 'Bad Request.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  400:
                    description: 'Bad request. Redirect to the same form.'
                    value: '/reset-password{token}'  

#M02 - Events
...
  /search:

    get:
      operationId: R202
      summary: 'R202: Search events'
      description: 'Search events based on name or description via an UI. Access: PUB'
      tags:
        - 'M02: Events'
      responses:
        '200':
          description: 'Ok. Show search results modal UI.'

    post:
      operationId: R203
      summary: 'R203: Search events'
      description: 'Search events based on name or description. Access: PUB'
      tags:
        - 'M02: Events'
      parameters:
        - in: path
          name: query_string
          schema:
            type: string
          required: true
      responses:
        '302':
          description: 'Redirect after processing the search query.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful search. Redirect to search results page.'
                  value: '/search'
                302Failure:
                  description: 'Failed search. Redirect to home page.'
                  value: '/home'
...
  /storeEvent:
    post:
        operationId: R206
        summary: 'R206: Create Event Action'
        description: 'This web resource allows the user to change its user profile information. Access: AUTH, OWN'
        tags:
          - 'M02: Events'

        requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  description:
                    type: string
                  date:
                    type: string
                    format: date
                  capacity:
                    type: integer
                  city:
                    type: string
                  country:
                    type: string
                  price:
                    type: number
                    format: float
                  address:
                    type: string
                  tag:
                    type: string
                  image:
                    type: string
                    format: binary
                required:
                  - name
                  - description
                  - date
                  - capacity
                  - city
                  - country
                  - price
                  - address
                  - tag

        responses:
          '302':
            description: 'Redirect after processing the new event information.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  302Success:
                    description: 'Successful. Redirect to new event page.'
                    value: '/event{id}'
                  302Failure:
                    description: 'Failed. Redirect to same create event page.'
                    value: '/createEvent'
          '400':
            description: 'Bad Request. Event not created.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  400:
                    description: 'Event not created. Redirect to the same create event page.'
                    value: '/createEvent'
...
  /deleteEvent:

    post:
        operationId: R209
        summary: 'R209: Delete Event Action'
        description: 'This web resource allows to delete an event. Access: AUTH, OWN, ADM'
        tags:
          - 'M02: Events'

        requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  eventid:
                    type: integer
                required:
                  - eventid
        responses:
          '302':
            description: 'Success'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  302Success:
                    description: 'Successful. Redirect to the home page.'
                    value: '/home'
                  302Failure:
                    description: 'Failed. Redirect to same event page.'
                    value: '/event{eventid}'
          '404':
            description: 'Not Found. Event not found.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  404:
                    description: 'Event not found. Redirect to the same event page.'
                    value: '/event{eventid}'
          '400':
            description: 'Bad Request. Event not deleted.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  400:
                    description: 'Event not deleted. Redirect to the same event page.'
                    value: '/event{eventid}'

#M03 - Event Participation and Ownership
...
  /addEventUsers:

      post:
        operationId: R303
        summary: 'R303: Add user to an event'
        description: 'Allows an event owner to add an user to an event. Access: AUTH, OWN'
        tags:
          - 'M03: Event Participation and Ownership'

        requestBody:
          required: true
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  eventid:
                    type: integer
                  userid:
                    type: integer
                required:
                  - eventid
                  - userid

        responses:
          '302':
            description: 'Success'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  302Success:
                    description: 'Successful. Redirect to the user page with the events managed.'
                    value: '/user{id}'
                  302Failure:
                    description: 'Failed. Redirect to the user page with the events managed.'
                    value: '/user{id}'
          '400':
            description: 'Bad Request.'
            headers:
              Location:
                schema:
                  type: string
                examples:
                  400:
                    description: 'Bad request. Redirect to the user page with the events managed.'
                    value: '/user{id}'
  /removeEventUsers:

    #exclusively used get in this resource for simplicity, as it is not a security risk

    post:
      operationId: R304
      summary: 'R304: Remove user from an event'
      description: 'Allows an event owner to remove an user from an event. Access: AUTH, OWN'
      tags:
        - 'M03: Event Participation and Ownership'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  eventid:
                    type: integer
                  userid:
                    type: integer
                required:
                  - eventid
                  - userid

      responses:
        '302':
          description: 'Success'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful. Redirect to the user page with the events managed.'
                  value: '/user{id}'
                302Failure:
                  description: 'Failed. Redirect to the user page with the events managed.'
                  value: '/user{id}'
        '400':
          description: 'Bad Request.'
          headers:
            Location:
              schema:
                type: string
              examples:
                400:
                  description: 'Bad request. Redirect to the user page with the events managed.'
                  value: '/user{id}'
  /transferOwnership:

    post:
      operationId: R305
      summary: 'R305: Transfer the ownership of an event'
      description: 'Allows an event owner to remove an tranfer the ownership of an event to another user. Access: OWN'
      tags:
        - 'M03: Event Participation and Ownership'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  eventid:
                    type: integer
                  userid:
                    type: integer
                required:
                  - eventid
                  - userid

      responses:
        '302':
          description: 'Success'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful. Redirect to the user of the event ehich ownership was transfered.'
                  value: '/event{eventid}'
                302Failure:
                  description: 'Failed.  Redirect to the page of the event ehich ownership was to be transfered.'
                  value: '/event{eventid}'
        '400':
          description: 'Bad Request.'
          headers:
            Location:
              schema:
                type: string
              examples:
                400:
                  description: 'Bad request. Redirect to the page of the event ehich ownership was to be transfered.'
                  value: '/event{eventid}'

#M04 - Review and Comment
  /storeComment:
    
        post:
          operationId: R401
          summary: 'R401: Comment event'
          description: 'Allows a user to comment an event. Access: AUTH'
          tags:
            - 'M04: Review and Comment'
          requestBody:
                    required: true
                    content:
                      application/x-www-form-urlencoded:
                        schema:
                          type: object
                          properties:
                            eventid:
                              type: integer
                            userid:
                              type: integer
                            comment:
                              type: string
                          required:
                            - eventid
                            - userid
                            - comment         
    
          responses:
            '302':
              description: 'Success'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    302Success:
                      description: 'Successful. Redirect to the event page.'
                      value: '/event{eventid}'
                    302Failure:
                      description: 'Failed. Redirect to same event page.'
                      value: '/event{eventid}'
            '404':
              description: 'Not Found. Event not found.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    404:
                      description: 'Event not found. Redirect to the home page.'
                      value: '/home'
            '400':
              description: 'Bad Request. Comment not processed.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    400:
                      description: 'Bad request. Comment not processed. Redirect to the same event page.'
                      value: '/event{eventid}'
  /deleteComment:
    
        post:
          operationId: R402
          summary: 'R402: Delete a comment from an event'
          description: 'Allows a user to delete a comment from an event. Access: AUTH, OWN'
          tags:
            - 'M04: Review and Comment'
          requestBody:
                    required: true
                    content:
                      application/x-www-form-urlencoded:
                        schema:
                          type: object
                          properties:
                            eventid:
                              type: integer
                            userid:
                              type: integer
                            comment:
                              type: string
                          required:
                            - commentid 
                            - userid  
                            - eventid        

          responses:
            '302':
              description: 'Success'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    302Success:
                      description: 'Successful. Redirect to the event page.'
                      value: '/event{eventid}'
                    302Failure:
                      description: 'Failed. Redirect to same event page.'
                      value: '/event{eventid}'
            '404':
              description: 'Not Found. Event not found.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    404:
                      description: 'Event not found. Redirect to the home page.'
                      value: '/home'
            '400':
              description: 'Bad Request. Could not delete comment.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    400:
                      description: 'Bad request. Could not delete comment. Redirect to the same event page.'
                      value: '/event{eventid}'
  /addUpvote:
    
        post:
          operationId: R403
          summary: 'R403: Upvote comment'
          description: 'Allows a user to upvote a comment in an event. Access: AUTH'
          tags:
            - 'M04: Review and Comment'
          requestBody:
                    required: true
                    content:
                      application/x-www-form-urlencoded:
                        schema:
                          type: object
                          properties:
                            commentid:
                              type: integer
                            userid:
                              type: integer
                          required:
                            - commentid
                            - userid      
    
          responses:
            '302':
              description: 'Success'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    302Success:
                      description: 'Successful. Redirect to the event page.'
                      value: '/event{eventid}'
                    302Failure:
                      description: 'Failed. Redirect to same event page.'
                      value: '/event{eventid}'
            '404':
              description: 'Not Found. Event not found.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    404:
                      description: 'Event not found. Redirect to the home page.'
                      value: '/home'
            '400':
              description: 'Bad Request. Upvote not processed.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    400:
                      description: 'Bad request. Upvote in comment not processed. Redirect to the same event page.'
                      value: '/event{eventid}'
...
  /getComments:
    
        post:
          operationId: R405
          summary: 'R405: Get all comments for an event'
          description: 'Retrieves all comments from an event. Access: AUTH'
          tags:
            - 'M04: Review and Comment'
          requestBody:
                    required: true
                    content:
                      application/x-www-form-urlencoded:
                        schema:
                          type: object
                          properties:
                            eventid:
                              type: integer
                          required:
                            - eventid      

          responses:
            '200':
              description: 'Success'
              content:
                application/json:
                  schema:
                    type: object
                    properties:
                      userid:
                        type: integer
                      username:
                        type: string
                      comment:
                        type: string
                      upvotes:
                        type: integer
            '404':
              description: 'Not Found. Event not found.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    404:
                      description: 'Event not found. Redirect to the home page.'
                      value: '/home'
            '400':
              description: 'Bad Request. Could not get comments.'
              headers:
                Location:
                  schema:
                    type: string
                  examples:
                    400:
                      description: 'Bad request. Could not get comments. Redirect to the home page'
                      value: '/home'
...